+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.4] - 2026-02-14
+
+### Added
+
+- Add timezone config for date filters.
+
+### Fixed
+
+- Fix updating table.
+- Skip invalid filter values and treat unknown filters as fulltext.
+
 ## [3.1.3] - 2026-02-11
 
 ### Fixed
@@ -551,6 +562,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[3.1.4]: https://https://github.com/internetguru/laravel-model-browser/compare/v3.1.3...v3.1.4
 [3.1.3]: https://https://github.com/internetguru/laravel-model-browser/compare/v3.1.2...v3.1.3
 [3.1.2]: https://https://github.com/internetguru/laravel-model-browser/compare/v3.1.1...v3.1.2
 [3.1.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v3.1.0...v3.1.1